split the lib into multiple files
