+++ Cargo.toml
@@ -28,7 +28,12 @@ log = "0.4"
 env_logger = "0.11"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
+pyo3-tch = { version = "0.20.0", optional = true }
 
 [dev-dependencies]
 hf-hub = "0.4.3"
 tqdm = "0.8.0"
+
+[features]
+default = []
+pytorch = [ "dep:pyo3-tch" ]
